Build results of dee934b (on master)

+++ t/tlib/Test/SpawnRedisServer.pm
@@ -9,17 +9,24 @@ use IPC::Cmd qw(can_run);
 use POSIX ":sys_wait_h";
 use base qw( Exporter );
 
-our @EXPORT = qw( redis );
+our @EXPORT    = qw( redis );
+our @EXPORT_OK = qw( redis reap );
+
+## FIXME: for the love of $Deity... move to Test::TCP, will you??
+my $port = 11011 + ($$ % 127);
 
 sub redis {
   my %params = (
     timeout => 120,
     @_,
   );
-  my ($fh, $fn) = File::Temp::tempfile();
-  my $port = 11011 + ($$ % 127);
 
-  unlink('redis-server.log');
+  my ($fh, $fn) = File::Temp::tempfile();
+
+  $port++;
+  my $addr = "127.0.0.1:$port";
+
+  unlink("redis-server-$addr.log");
   unlink('dump.rdb');
 
   $fh->print("
@@ -29,15 +36,14 @@ sub redis {
     port $port
     bind 127.0.0.1
     loglevel debug
-    logfile redis-server.log
+    logfile redis-server-$addr.log
   ");
   $fh->flush;
 
-  my $addr = "127.0.0.1:$port";
   Test::More::diag("Spawn Redis at $addr, cfg $fn") if $ENV{REDIS_DEBUG};
 
   my $redis_server_path = $ENV{REDIS_SERVER_PATH} || 'redis-server';
-  if (! can_run($redis_server_path)) {
+  if (!can_run($redis_server_path)) {
     Test::More::plan skip_all => "Could not find binary redis-server";
     return;
   }
@@ -78,10 +84,11 @@ sub spawn_server {
 
     my $redis   = Redis->new(server => $addr, reconnect => 5, every => 200);
     my $version = $redis->info->{redis_version};
-    my $alive   = 1;
+    my $alive   = $$;
 
     my $c = sub {
       return unless $alive;
+      return unless $$ == $alive;    ## only our creator can kill us
 
       Test::More::diag("Killing server at $pid") if $ENV{REDIS_DEBUG};
       kill(15, $pid);
@@ -89,9 +96,11 @@ sub spawn_server {
       my $failed = reap($pid);
       Test::More::diag("Failed to kill server at $pid")
         if $ENV{REDIS_DEBUG} and $failed;
-      unlink('redis-server.log');
+      unlink("redis-server-$addr.log");
       unlink('dump.rdb');
       $alive = 0;
+
+      return !$failed;
     };
 
     return $version => $c;
@@ -106,11 +115,12 @@ sub spawn_server {
 }
 
 sub reap {
-  my ($pid) = @_;
-  $pid = -1 unless $pid;
+  my ($pid, $limit) = @_;
+  $pid   = -1 unless $pid;
+  $limit = 3  unless $limit;
 
   my $try = 0;
-  while ($try++ < 3) {
+  while ($try++ < $limit) {
     my $ok = waitpid($pid, WNOHANG);
     $try = 0, last if $ok > 0;
     sleep(1);
